--- src/app/api/fetchCourseCat/route.ts
@@ -1,7 +1,0 @@
-import { main } from "@/services/courseService";
-
-export async function GET() {
-  const data = await main();
-
-  return Response.json(data);
-}
